Move ltr.feature.norm classes into their own ltr.norm package (since now Feature.java now longer uses Normalizer.java).
