+++ solr/contrib/ltr/src/java/org/apache/solr/ltr/log/LoggingModel.java
@@ -22,8 +22,8 @@ import java.util.Map;
 
 import org.apache.lucene.index.LeafReaderContext;
 import org.apache.lucene.search.Explanation;
-import org.apache.solr.ltr.feature.norm.Normalizer;
 import org.apache.solr.ltr.model.LTRScoringModel;
+import org.apache.solr.ltr.norm.Normalizer;
 import org.apache.solr.ltr.ranking.Feature;
 import org.apache.solr.ltr.util.LTRUtils;
