+++ solr/contrib/ltr/src/java/org/apache/solr/ltr/norm/Normalizer.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.solr.ltr.feature.norm;
+package org.apache.solr.ltr.norm;
 
 import java.util.LinkedHashMap;
 import java.util.Map;
