+++ solr/contrib/ltr/src/java/org/apache/solr/ltr/norm/impl/MinMaxNormalizer.java
@@ -14,10 +14,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.solr.ltr.feature.norm.impl;
+package org.apache.solr.ltr.norm.impl;
 
 import java.util.LinkedHashMap;
-import org.apache.solr.ltr.feature.norm.Normalizer;
+
+import org.apache.solr.ltr.norm.Normalizer;
 
 public class MinMaxNormalizer extends Normalizer {
