+++ solr/contrib/ltr/src/java/org/apache/solr/ltr/norm/impl/StandardNormalizer.java
@@ -14,11 +14,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.solr.ltr.feature.norm.impl;
+package org.apache.solr.ltr.norm.impl;
 
 import java.util.LinkedHashMap;
 
-import org.apache.solr.ltr.feature.norm.Normalizer;
+import org.apache.solr.ltr.norm.Normalizer;
 
 public class StandardNormalizer extends Normalizer {
