+++ solr/contrib/ltr/src/java/org/apache/solr/ltr/norm/impl/package-info.java
@@ -14,33 +14,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.solr.ltr.feature.norm.impl;
 
-import java.util.LinkedHashMap;
-
-import org.apache.solr.ltr.feature.norm.Normalizer;
-
-public class IdentityNormalizer extends Normalizer {
-
-  public static final IdentityNormalizer INSTANCE = new IdentityNormalizer();
-
-  public IdentityNormalizer() {
-
-  }
-
-  @Override
-  public float normalize(float value) {
-    return value;
-  }
-
-  @Override
-  public LinkedHashMap<String,Object> paramsToMap() {
-    return null;
-  }
-
-  @Override
-  public String toString() {
-    return getClass().getSimpleName();
-  }
-
-}
+/**
+ * This package contains several implementations of normalizers.
+ */
+package org.apache.solr.ltr.norm.impl;
