+++ solr/contrib/ltr/src/java/org/apache/solr/ltr/ranking/RankSVMModel.java
@@ -22,9 +22,9 @@ import java.util.Map;
 
 import org.apache.lucene.index.LeafReaderContext;
 import org.apache.lucene.search.Explanation;
-import org.apache.solr.ltr.feature.norm.Normalizer;
 import org.apache.solr.ltr.model.LTRScoringModel;
 import org.apache.solr.ltr.model.ModelException;
+import org.apache.solr.ltr.norm.Normalizer;
 
 public class RankSVMModel extends LTRScoringModel {
