+++ solr/contrib/ltr/src/test-files/modelExamples/ranksvm-model.json
@@ -8,7 +8,7 @@
         {
             "name":"popularity",
             "norm": {
-                "class":"org.apache.solr.ltr.feature.norm.impl.MinMaxNormalizer",
+                "class":"org.apache.solr.ltr.norm.impl.MinMaxNormalizer",
                 "params":{ "min":"0.0f", "max":"10.0f" }
             }
         },
