+++ solr/contrib/ltr/src/test/org/apache/solr/ltr/norm/impl/TestStandardNormalizer.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.solr.ltr.feature.norm.impl;
+package org.apache.solr.ltr.norm.impl;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
@@ -23,7 +23,8 @@ import java.util.HashMap;
 import java.util.Map;
 
 import org.apache.solr.core.SolrResourceLoader;
-import org.apache.solr.ltr.feature.norm.Normalizer;
+import org.apache.solr.ltr.norm.Normalizer;
+import org.apache.solr.ltr.norm.impl.StandardNormalizer;
 import org.junit.Test;
 
 public class TestStandardNormalizer {
